refactor(trending-item): enhance card layout with responsive grid, hover effect, and styled rating badge
