+++ src/Components/Movies/Movies.jsx
@@ -1,7 +1,7 @@
 import { useTrending } from "../Hooks/useTrending";
 import Loading from "../Loading/Loading";
 import TrendingHeader from "../Ui/TrendingHeader";
-import TrendingItem from "../Ui/TrendingItem";
+import TrendingItem from "../Ui/TrendingItem/TrendingItem";
 
 export default function Movies() {
   const { isLoading, trendingList } = useTrending("movie");
