+++ src/Components/People/People.jsx
@@ -1,7 +1,7 @@
 import { useTrending } from "../Hooks/useTrending";
 import Loading from "../Loading/Loading";
 import TrendingHeader from "../Ui/TrendingHeader";
-import TrendingItem from "../Ui/TrendingItem";
+import TrendingItem from "../Ui/TrendingItem/TrendingItem";
 
 export default function People() {
   const { isLoading, trendingList } = useTrending("person");
