+++ src/Components/Ui/TrendingItem/TrendingItem.module.css
@@ -1,0 +1,15 @@
+.card-hover {
+  transition: transform 0.3s ease, box-shadow 0.3s ease;
+  cursor: pointer;
+}
+
+.card-hover:hover {
+  transform: scale(1.05);
+  transition: transform 0.3s ease, box-shadow 0.3s ease;
+  box-shadow: 0 8px 20px rgba(0, 0, 0, 0.4);
+  cursor: pointer;
+}
+
+.badg-shadow {
+  box-shadow: 0 6px 20px rgba(0, 0, 0,0.4);
+}
