+++ src/index.css
@@ -1,20 +1,21 @@
-@import url('https://fonts.googleapis.com/css2?family=Montserrat+Alternates:wght@300&display=swap');
+@import url("https://fonts.googleapis.com/css2?family=Montserrat+Alternates:wght@300&display=swap");
 
-body{
+body {
   background-color: #131722;
   color: #fff !important;
   font-family: "Montserrat Alternates", sans-serif;
 }
 
-ul,li{
+ul,
+li {
   list-style-type: none;
 }
 
-.main-bg{
-  background-color:  #24baef;
+.main-bg {
+  background-color: #24baef;
 }
 
-.brdr{
+.brdr {
   height: 2px;
   background-color: #a8a8a83d;
 }
