+++ src/index.js
@@ -2,7 +2,6 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import "bootstrap/dist/css/bootstrap.css";
 import "bootstrap/dist/js/bootstrap.bundle.js";
-import "../node_modules/@fortawesome/fontawesome-free/css/all.min.css"
 import "./index.css";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
